+++ eyes.sdk.core/src/main/java/com/applitools/eyes/fluent/BatchClose.java
@@ -1,6 +1,7 @@
 package com.applitools.eyes.fluent;
 
 import com.applitools.eyes.Logger;
+import com.applitools.eyes.ProxySettings;
 import com.applitools.utils.ArgumentGuard;
 import com.applitools.utils.GeneralUtils;
 
@@ -9,6 +10,8 @@ import java.util.List;
 public class BatchClose {
     protected final Logger logger;
     protected String serverUrl;
+    protected String apiKey;
+    protected ProxySettings proxySettings;
 
     public BatchClose() {
         this(new Logger());
@@ -20,13 +23,35 @@ public class BatchClose {
     }
 
     public BatchClose setUrl(String url) {
+        ArgumentGuard.notNull(url, "url");
         serverUrl = url;
+        return this;
+    }
+
+    public BatchClose setApiKey(String apiKey) {
+        ArgumentGuard.notNull(apiKey, "apiKey");
+        this.apiKey = apiKey;
+        return this;
+    }
+
+    public BatchClose setProxy(ProxySettings proxySettings) {
+        ArgumentGuard.notNull(proxySettings, "proxySettings");
+        this.proxySettings = proxySettings;
         return this;
     }
 
     public EnabledBatchClose setBatchId(List<String> batchIds) {
         ArgumentGuard.notNull(batchIds, "batchIds");
         ArgumentGuard.notContainsNull(batchIds, "batchIds");
-        return new EnabledBatchClose(logger, serverUrl, batchIds);
+        EnabledBatchClose enabledBatchClose = new EnabledBatchClose(logger, serverUrl, batchIds);
+        if (apiKey != null) {
+            enabledBatchClose.setApiKey(apiKey);
+        }
+
+        if (proxySettings != null) {
+            enabledBatchClose.setProxy(proxySettings);
+        }
+
+        return enabledBatchClose;
     }
 }