+++ eyes.sdk.core/src/main/java/com/applitools/eyes/fluent/EnabledBatchClose.java
@@ -2,6 +2,7 @@ package com.applitools.eyes.fluent;
 
 import com.applitools.connectivity.ServerConnector;
 import com.applitools.eyes.Logger;
+import com.applitools.eyes.ProxySettings;
 import com.applitools.utils.ArgumentGuard;
 
 import java.util.List;
@@ -18,13 +19,30 @@ public class EnabledBatchClose extends BatchClose {
     }
 
     @Override
-    public com.applitools.eyes.fluent.EnabledBatchClose setUrl(String url) {
+    public EnabledBatchClose setUrl(String url) {
+        ArgumentGuard.notNull(url, "url");
         serverUrl = url;
         return this;
     }
 
     @Override
-    public com.applitools.eyes.fluent.EnabledBatchClose setBatchId(List<String> batchIds) {
+    public EnabledBatchClose setApiKey(String apiKey) {
+        ArgumentGuard.notNull(apiKey, "apiKey");
+        this.apiKey = apiKey;
+        serverConnector.setApiKey(apiKey);
+        return this;
+    }
+
+    @Override
+    public EnabledBatchClose setProxy(ProxySettings proxySettings) {
+        ArgumentGuard.notNull(proxySettings, "proxySettings");
+        this.proxySettings = proxySettings;
+        serverConnector.setProxy(proxySettings);
+        return this;
+    }
+
+    @Override
+    public EnabledBatchClose setBatchId(List<String> batchIds) {
         ArgumentGuard.notNull(batchIds, "batchIds");
         ArgumentGuard.notContainsNull(batchIds, "batchIds");
         this.batchIds = batchIds;